Rejoin ast walking components
Breaking them into separate modules (See: f69b39) was nice in theory but it ends up being more of a circular dependency nightmare than it's worth
