--- .coveragerc
@@ -1,2 +1,0 @@
-[run]
-parallel = True
