+++ flake8_annotations/checker.py
@@ -4,14 +4,14 @@ import typing as t
 from functools import lru_cache
 
 from flake8_annotations import PY_GTE_38, __version__, enums, error_codes
-from flake8_annotations.visitors import FunctionVisitor, ast
+from flake8_annotations.ast_walker import FunctionVisitor, ast
 
 if t.TYPE_CHECKING:
     from argparse import Namespace
 
     from flake8.options.manager import OptionManager
 
-    from flake8_annotations.models import Argument, Function
+    from flake8_annotations.ast_walker import Argument, Function
 
 FORMATTED_ERROR = t.Tuple[int, int, str, t.Type[t.Any]]
