+++ testing/test_classifier.py
@@ -3,10 +3,10 @@ from typing import Tuple
 import pytest
 import pytest_check as check
 
+from flake8_annotations.ast_walker import Argument, Function
 from flake8_annotations.checker import FORMATTED_ERROR, classify_error
 from flake8_annotations.enums import AnnotationType
 from flake8_annotations.error_codes import Error
-from flake8_annotations.models import Argument, Function
 from testing.helpers import check_source
 from testing.test_cases import classifier_object_attributes
 from testing.test_cases.type_comment_test_cases import ParserTestCase, parser_test_cases
