+++ tox.ini
@@ -25,7 +25,7 @@ exclude=
     error_codes.py,
     .tox
 per-file-ignores =
-    testing/test_*.py:TC
+    testing/test_*.py:TC E501 D103,
 import-order-style=pycharm
 application-import-names=flake8_annotations,testing
